backend/requirements.txt
asgiref==3.8.1
boto3==1.35.49
botocore==1.35.49
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
dj-database-url==2.3.0
Django==5.1.2
django-anymail==12.0
django-ckeditor-5==0.2.15
django-cors-headers==4.5.0
django-jazzmin==3.0.1
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
environs==11.0.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
marshmallow==3.23.0
packaging==24.1
pillow==11.0.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.3
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1
stripe==11.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
